tools/elfutils: add maintainer-clean to clean target
Use the clean recipe during build, so that files already
generated by the bootstrapping for the release are removed
before the bootstrap script is run again.

The configure stage must be run twice to both clean and build,
because maintainer-clean will delete the Makefile.

Building from tracked sources only requires a lexical analyzer,
so add dependency for flex, and enable maintainer mode.
A header issue related to use of gnulib with the lex source is patched.

Signed-off-by: Michael Pratt <mcpratt@pm.me>